requirements.txt
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.0
cryptography==42.0.8
et-xmlfile==1.1.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==2.1.5
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
ply==3.11
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pysmi==0.3.4
pysmi-lextudio==1.4.3
pysnmp-lextudio==6.2.0
pysnmpcrypto==0.0.4
PySocks==1.7.1
pytest==8.2.2
pytest-cov==4.1.0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.19
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
